fix circular import, temporary location for merge_headers
