+++ tfs/frame.py
@@ -6,15 +6,15 @@ Contains the class definition of a ``TfsDataFrame``, inherited from the ``pandas
 as a utility function to validate the correctness of a ``TfsDataFrame``.
 """
 import logging
+from collections import OrderedDict
 from contextlib import suppress
 from functools import partial, reduce
-from typing import Optional, Sequence, Union
+from typing import Sequence, Union
 
 import numpy as np
 import pandas as pd
 
 from tfs.errors import TfsFormatError
-from tfs.tools import merge_headers
 
 LOGGER = logging.getLogger(__name__)
 
@@ -177,6 +177,33 @@ class TfsDataFrame(pd.DataFrame):
             else merge_headers(self.headers, right.headers, how=how_headers)
         )
         return TfsDataFrame(data=dframe, headers=new_headers)
+
+
+def merge_headers(headers_left: dict, headers_right: dict, how: str) -> OrderedDict:
+    """
+    Merge new_headers of two ``TfsDataFrames`` together.
+
+    Args:
+        headers_left (dict): TODO.
+        headers_right (dict): TODO.
+        how (str): Type of merge to be performed, either **left** or **right**. If **left*, prioritize keys
+            from **headers_left** in case of duplicate keys. If **right**, prioritize keys from
+            **headers_right** in case of duplicate keys.
+
+    Returns:
+        A new ``OrderedDict`` as the merge of the two provided dictionaries.
+    """
+    accepted_merges: Set[str] = {"left", "right"}
+    if how not in accepted_merges:
+        raise ValueError(f"Invalid 'how' argument, should be one of {accepted_merges}")
+    LOGGER.debug(f"Merging new_headers with method '{how}'")
+    if how == "left":  # we prioritize the contents of headers_left
+        result = headers_right.copy()
+        result.update(headers_left)
+    else:  # we prioritize the contents of headers_right
+        result = headers_left.copy()
+        result.update(headers_right)
+    return OrderedDict(result)  # so that the TfsDataFrame still has an OrderedDict as header
 
 
 def concat(
